fix: remove old delegates before writing new ones (#32)
Because we do not write all delegates, only those that have some votes
if delegate has all of its delegations removed it would not be reflected
by the indexer.

To avoid this before we update delegates that we know have some score
we delete all other delegates - this means all delegates that lost VP
will be removed from the list.